Add api call for single movie, add err page, organize folders
