+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/index.js":"1","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/reportWebVitals.js":"2","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/App.js":"3","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/movieData.js":"4","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Movies.js":"5","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MoviePreview.js":"6","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Navbar.js":"7","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MovieCard.js":"8","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/apiCalls.js":"9"},{"size":500,"mtime":1606861541033,"results":"10","hashOfConfig":"11"},{"size":362,"mtime":1606861541034,"results":"12","hashOfConfig":"11"},{"size":1653,"mtime":1607044954214,"results":"13","hashOfConfig":"11"},{"size":12967,"mtime":1606962709593,"results":"14","hashOfConfig":"11"},{"size":524,"mtime":1607038002552,"results":"15","hashOfConfig":"11"},{"size":1385,"mtime":1607044954216,"results":"16","hashOfConfig":"11"},{"size":237,"mtime":1607035551599,"results":"17","hashOfConfig":"11"},{"size":373,"mtime":1607035551596,"results":"18","hashOfConfig":"11"},{"size":343,"mtime":1607044954218,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"1tg657g",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"35","usedDeprecatedRules":"22"},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/index.js",[],["40","41"],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/reportWebVitals.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/App.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/movieData.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Movies.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MoviePreview.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Navbar.js",["42"],"import React, { Component } from 'react';\nimport './Navbar.scss'\n\nconst Navbar = () =>  {\n\n    return (\n      <nav className='navbar'>\n        <h1 className='site-title'>Rancid Tomatillos</h1>\n      </nav>\n    )\n}\n\nexport default Navbar;","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MovieCard.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/apiCalls.js",[],{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","severity":1,"message":"48","line":1,"column":17,"nodeType":"49","messageId":"50","endLine":1,"endColumn":26},"no-native-reassign",["51"],"no-negated-in-lhs",["52"],"no-unused-vars","'Component' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
+[{"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/index.js":"1","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/reportWebVitals.js":"2","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/apiCalls.js":"3","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MovieCard/MovieCard.js":"4","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Movies/Movies.js":"5","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MoviePreview/MoviePreview.js":"6","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Navbar/Navbar.js":"7","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/ErrorPage/ErrorPage.js":"8","/Users/shawn/turing/mod3/Rotten-Tomatillos/src/App.js":"9"},{"size":500,"mtime":1607134580160,"results":"10","hashOfConfig":"11"},{"size":362,"mtime":1606861541034,"results":"12","hashOfConfig":"11"},{"size":379,"mtime":1607133338811,"results":"13","hashOfConfig":"11"},{"size":373,"mtime":1607134045951,"results":"14","hashOfConfig":"11"},{"size":535,"mtime":1607134400785,"results":"15","hashOfConfig":"11"},{"size":1385,"mtime":1607134343424,"results":"16","hashOfConfig":"11"},{"size":222,"mtime":1607134435608,"results":"17","hashOfConfig":"11"},{"size":308,"mtime":1607134300156,"results":"18","hashOfConfig":"11"},{"size":1853,"mtime":1607134687862,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1tg657g",{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/index.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/reportWebVitals.js",[],["39","40"],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/apiCalls.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MovieCard/MovieCard.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Movies/Movies.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/MoviePreview/MoviePreview.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/Navbar/Navbar.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/ErrorPage/ErrorPage.js",[],"/Users/shawn/turing/mod3/Rotten-Tomatillos/src/App.js",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
