+++ src/App.scss
@@ -8,6 +8,7 @@
 html {
   height: 100%;
   width: 100%;
+  background-color: $dark-blue;
 }
 
 .app {
