+++ src/ErrorPage/ErrorPage.scss
@@ -1,0 +1,12 @@
+@import '../variables.scss';
+
+.error-page {
+  height: 100vh;
+  color: $white;
+  width: 100%;
+  position: fixed;
+  display: flex;
+  flex-direction: column;
+  color: #FDFFFC;
+  justify-content: center;
+}
