+++ src/MovieCard/MovieCard.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .movie-card {
   height: 20em;
