+++ src/MoviePreview/MoviePreview.js
