+++ src/MoviePreview/MoviePreview.scss
@@ -1,4 +1,4 @@
-@import './variables.scss';
+@import '../variables.scss';
 
 .movie-preview-page {
   height: 100vh;
