+++ src/Movies/Movies.scss
@@ -1,4 +1,4 @@
-@import './variables.scss';
+@import '../variables.scss';
 
 .movies-list {
   display: flex;
