+++ src/Navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@import './_variables.scss';
+@import '../_variables.scss';
 
 .navbar {
   border-bottom: $light-blue solid 2px;
