+++ src/apiCalls.js
@@ -5,5 +5,6 @@ export const getAllMovieData = async () => {
 
 export const getSingleMovieData = async (id) => {
   const response = await fetch(`https://rancid-tomatillos.herokuapp.com/api/v2/movies/${id}`)
+  console.log(response.statusText);
   return await response.json()
 }
